fix submitTest to testinput and test online, response code after submit test

+++ src/main/java/com/englishacademy/EnglishAcademy/dtos/testInputStudent/TestInputStudentDTO.java
@@ -24,7 +24,7 @@ public class TestInputStudentDTO {
     private Integer totalQuestionVocabulary;
     private Integer totalQuestionGrammar;
     private Double score;
-    private Double time;
+    private Integer time;
     private Timestamp createdDate;
     private Timestamp modifiedDate;
     private String createdBy;
